Skip GitHub API checks when the data can't be fetched ARCHBOM-1310 (#36)
Get around the issue fetching data from GitHub for edx-blog (and any similar problems that may crop up later in the repo health dashboard job). I'd like to enhance this later to have the fixture report the exact problem encountered without blowing up the test suite, but for now I just want to get the new data in the dashboard for most of our repos.