Add ownership checks ARCHBOM-1383 (#44)
Added a new check to get repository ownership data from the master spreadsheet, and added `gspread` as a dependency. The new check should be skipped unless the appropriate environment variables are set (which will only be true for the Jenkins job and some local dev environments).

It looks like some of the requirements files (like `travis.txt`) have too much in them, but I'll ticket fixing that separately.